Back/requirements.txt
flask
pyrebase
flask_cors
pyopenssl
Babel
Money